Add files via upload
Write a program to determine the ranges of char, short, int, and long variables, both signed and unsigned, by printing appropriate values from standard headers and by direct computation. Harder if you compute them: determine the ranges of the various floating-point types.
The program does not include the computation of MAX float and double, because at 16777216, the mantis can't include 16777217.